# artifacts/api-server/src/routes/tasks.ts
import { Router, type IRouter } from "express";
import { eq, inArray, sql } from "drizzle-orm";
import { db, tasksTable, taskAssigneesTable, usersTable } from "@workspace/db";
import {
  CreateTaskBody,
  UpdateTaskBody,
  GetTaskParams,
  UpdateTaskParams,
  DeleteTaskParams,
  AddFeedbackParams,
  AddFeedbackBody,
} from "@workspace/api-zod";
import { requireAuth, requireManager } from "../lib/auth";

const router: IRouter = Router();

async function getTaskWithAssignees(taskId: number) {
  const [task] = await db.select().from(tasksTable).where(eq(tasksTable.id, taskId));
  if (!task) return null;

  const assigneeRows = await db
    .select({
      id: usersTable.id,
      email: usersTable.email,
      name: usersTable.name,
      role: usersTable.role,
      createdAt: usersTable.createdAt,
    })
    .from(taskAssigneesTable)
    .innerJoin(usersTable, eq(taskAssigneesTable.userId, usersTable.id))
    .where(eq(taskAssigneesTable.taskId, taskId));

  return {
    ...task,
    deadline: task.deadline ?? null,
    completionPercent: task.completionPercent ?? null,
    expectedCompletionTime: task.expectedCompletionTime ?? null,
    feedback: task.feedback ?? null,
    createdAt: task.createdAt.toISOString(),
    updatedAt: task.updatedAt.toISOString(),
    assignees: assigneeRows.map(u => ({ ...u, createdAt: u.createdAt.toISOString() })),
  };
}

router.get("/tasks/summary", requireAuth, async (req, res): Promise<void> => {
  let whereClause = undefined;

  if (req.session.role === "employee") {
    const assignedRows = await db
      .select({ taskId: taskAssigneesTable.taskId })
      .from(taskAssigneesTable)
      .where(eq(taskAssigneesTable.userId, req.session.userId!));
    const assignedIds = assignedRows.map(r => r.taskId);

    if (assignedIds.length === 0) {
      res.json({ todo: 0, in_progress: 0, done: 0, total: 0 });
      return;
    }
    whereClause = inArray(tasksTable.id, assignedIds);
  }

  const rows = await db
    .select({ status: tasksTable.status, count: sql<number>`count(*)::int` })
    .from(tasksTable)
    .where(whereClause)
    .groupBy(tasksTable.status);

  const summary = { todo: 0, in_progress: 0, done: 0, total: 0 };
  for (const row of rows) {
    const key = row.status as keyof typeof summary;
    if (key in summary) {
      summary[key] = row.count;
      summary.total += row.count;
    }
  }

  res.json(summary);
});

router.get("/tasks", requireAuth, async (req, res): Promise<void> => {
  let taskIds: number[] | null = null;

  if (req.session.role === "employee") {
    const assignedRows = await db
      .select({ taskId: taskAssigneesTable.taskId })
      .from(taskAssigneesTable)
      .where(eq(taskAssigneesTable.userId, req.session.userId!));
    taskIds = assignedRows.map(r => r.taskId);
  }

  let tasks;
  if (taskIds !== null) {
    if (taskIds.length === 0) {
      res.json([]);
      return;
    }
    tasks = await db.select().from(tasksTable).where(inArray(tasksTable.id, taskIds));
  } else {
    tasks = await db.select().from(tasksTable);
  }

  const result = await Promise.all(tasks.map(t => getTaskWithAssignees(t.id)));
  res.json(result.filter(Boolean));
});

router.post("/tasks", requireManager, async (req, res): Promise<void> => {
  const parsed = CreateTaskBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { assigneeIds, ...taskData } = parsed.data;

  const [task] = await db.insert(tasksTable).values({
    ...taskData,
    createdById: req.session.userId!,
  }).returning();

  if (assigneeIds && assigneeIds.length > 0) {
    await db.insert(taskAssigneesTable).values(
      assigneeIds.map(userId => ({ taskId: task.id, userId }))
    );
  }

  const result = await getTaskWithAssignees(task.id);
  res.status(201).json(result);
});

router.get("/tasks/:id", requireAuth, async (req, res): Promise<void> => {
  const params = GetTaskParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid task ID" });
    return;
  }

  const task = await getTaskWithAssignees(params.data.id);
  if (!task) {
    res.status(404).json({ error: "Task not found" });
    return;
  }

  res.json(task);
});

router.patch("/tasks/:id", requireAuth, async (req, res): Promise<void> => {
  const params = UpdateTaskParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid task ID" });
    return;
  }

  const parsed = UpdateTaskBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { assigneeIds, ...taskData } = parsed.data;

  const [existing] = await db.select().from(tasksTable).where(eq(tasksTable.id, params.data.id));
  if (!existing) {
    res.status(404).json({ error: "Task not found" });
    return;
  }

  if (Object.keys(taskData).length > 0) {
    await db.update(tasksTable).set(taskData).where(eq(tasksTable.id, params.data.id));
  }

  if (assigneeIds !== undefined) {
    await db.delete(taskAssigneesTable).where(eq(taskAssigneesTable.taskId, params.data.id));
    if (assigneeIds.length > 0) {
      await db.insert(taskAssigneesTable).values(
        assigneeIds.map(userId => ({ taskId: params.data.id, userId }))
      );
    }
  }

  const result = await getTaskWithAssignees(params.data.id);
  res.json(result);
});

router.delete("/tasks/:id", requireManager, async (req, res): Promise<void> => {
  const params = DeleteTaskParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid task ID" });
    return;
  }

  const [deleted] = await db.delete(tasksTable).where(eq(tasksTable.id, params.data.id)).returning();
  if (!deleted) {
    res.status(404).json({ error: "Task not found" });
    return;
  }

  res.sendStatus(204);
});

router.post("/tasks/:id/feedback", requireManager, async (req, res): Promise<void> => {
  const params = AddFeedbackParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: "Invalid task ID" });
    return;
  }

  const parsed = AddFeedbackBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [existing] = await db.select().from(tasksTable).where(eq(tasksTable.id, params.data.id));
  if (!existing) {
    res.status(404).json({ error: "Task not found" });
    return;
  }

  await db.update(tasksTable).set({ feedback: parsed.data.feedback }).where(eq(tasksTable.id, params.data.id));

  const result = await getTaskWithAssignees(params.data.id);
  res.json(result);
});

export default router;