# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  manager: "manager",
  employee: "employee",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface CreateTaskBody {
  title: string;
  description: string;
  /** @nullable */
  deadline?: string | null;
  assigneeIds?: number[];
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface UpdateTaskBody {
  status?: UpdateTaskBodyStatus;
  /** @nullable */
  completionPercent?: number | null;
  /** @nullable */
  expectedCompletionTime?: string | null;
  title?: string;
  description?: string;
  /** @nullable */
  deadline?: string | null;
  assigneeIds?: number[];
}

export interface FeedbackBody {
  feedback: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface Task {
  id: number;
  title: string;
  description: string;
  status: TaskStatus;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  completionPercent?: number | null;
  /** @nullable */
  expectedCompletionTime?: string | null;
  /** @nullable */
  feedback?: string | null;
  createdById: number;
  createdAt: string;
  updatedAt: string;
}

export type TaskWithAssigneesStatus =
  (typeof TaskWithAssigneesStatus)[keyof typeof TaskWithAssigneesStatus];

export const TaskWithAssigneesStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface TaskWithAssignees {
  id: number;
  title: string;
  description: string;
  status: TaskWithAssigneesStatus;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  completionPercent?: number | null;
  /** @nullable */
  expectedCompletionTime?: string | null;
  /** @nullable */
  feedback?: string | null;
  createdById: number;
  createdAt: string;
  updatedAt: string;
  assignees: User[];
}

export interface TaskSummary {
  todo: number;
  in_progress: number;
  done: number;
  total: number;
}