# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Task Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: tasks
    description: Task management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (manager only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (managers see all, employees see assigned)
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskWithAssignees"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task (manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithAssignees"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a single task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithAssignees"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithAssignees"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task (manager only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /tasks/{id}/feedback:
    post:
      operationId: addFeedback
      tags: [tasks]
      summary: Add manager feedback to a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithAssignees"

  /tasks/summary:
    get:
      operationId: getTaskSummary
      tags: [tasks]
      summary: Get task count summary grouped by status
      responses:
        "200":
          description: Task summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [manager, employee]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        deadline:
          type: ["string", "null"]
        assigneeIds:
          type: array
          items:
            type: integer
      required:
        - title
        - description

    UpdateTaskBody:
      type: object
      properties:
        status:
          type: string
          enum: [todo, in_progress, done]
        completionPercent:
          type: ["integer", "null"]
        expectedCompletionTime:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        deadline:
          type: ["string", "null"]
        assigneeIds:
          type: array
          items:
            type: integer

    FeedbackBody:
      type: object
      properties:
        feedback:
          type: string
      required:
        - feedback

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        deadline:
          type: ["string", "null"]
        completionPercent:
          type: ["integer", "null"]
        expectedCompletionTime:
          type: ["string", "null"]
        feedback:
          type: ["string", "null"]
        createdById:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - description
        - status
        - createdById
        - createdAt
        - updatedAt

    TaskWithAssignees:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        deadline:
          type: ["string", "null"]
        completionPercent:
          type: ["integer", "null"]
        expectedCompletionTime:
          type: ["string", "null"]
        feedback:
          type: ["string", "null"]
        createdById:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required:
        - id
        - title
        - description
        - status
        - createdById
        - createdAt
        - updatedAt
        - assignees

    TaskSummary:
      type: object
      properties:
        todo:
          type: integer
        in_progress:
          type: integer
        done:
          type: integer
        total:
          type: integer
      required:
        - todo
        - in_progress
        - done
        - total