# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with email and password
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
    role: zod.enum(["manager", "employee"]),
    createdAt: zod.string(),
  }),
  message: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  role: zod.enum(["manager", "employee"]),
  createdAt: zod.string(),
});

/**
 * @summary List all users (manager only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  role: zod.enum(["manager", "employee"]),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary List tasks (managers see all, employees see assigned)
 */
export const ListTasksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  status: zod.enum(["todo", "in_progress", "done"]),
  deadline: zod.string().nullish(),
  completionPercent: zod.number().nullish(),
  expectedCompletionTime: zod.string().nullish(),
  feedback: zod.string().nullish(),
  createdById: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  assignees: zod.array(
    zod.object({
      id: zod.number(),
      email: zod.string(),
      name: zod.string(),
      role: zod.enum(["manager", "employee"]),
      createdAt: zod.string(),
    }),
  ),
});
export const ListTasksResponse = zod.array(ListTasksResponseItem);

/**
 * @summary Create a task (manager only)
 */
export const CreateTaskBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  deadline: zod.string().nullish(),
  assigneeIds: zod.array(zod.number()).optional(),
});

/**
 * @summary Get a single task
 */
export const GetTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  status: zod.enum(["todo", "in_progress", "done"]),
  deadline: zod.string().nullish(),
  completionPercent: zod.number().nullish(),
  expectedCompletionTime: zod.string().nullish(),
  feedback: zod.string().nullish(),
  createdById: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  assignees: zod.array(
    zod.object({
      id: zod.number(),
      email: zod.string(),
      name: zod.string(),
      role: zod.enum(["manager", "employee"]),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Update a task
 */
export const UpdateTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTaskBody = zod.object({
  status: zod.enum(["todo", "in_progress", "done"]).optional(),
  completionPercent: zod.number().nullish(),
  expectedCompletionTime: zod.string().nullish(),
  title: zod.string().optional(),
  description: zod.string().optional(),
  deadline: zod.string().nullish(),
  assigneeIds: zod.array(zod.number()).optional(),
});

export const UpdateTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  status: zod.enum(["todo", "in_progress", "done"]),
  deadline: zod.string().nullish(),
  completionPercent: zod.number().nullish(),
  expectedCompletionTime: zod.string().nullish(),
  feedback: zod.string().nullish(),
  createdById: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  assignees: zod.array(
    zod.object({
      id: zod.number(),
      email: zod.string(),
      name: zod.string(),
      role: zod.enum(["manager", "employee"]),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Delete a task (manager only)
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Add manager feedback to a task
 */
export const AddFeedbackParams = zod.object({
  id: zod.coerce.number(),
});

export const AddFeedbackBody = zod.object({
  feedback: zod.string(),
});

export const AddFeedbackResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  status: zod.enum(["todo", "in_progress", "done"]),
  deadline: zod.string().nullish(),
  completionPercent: zod.number().nullish(),
  expectedCompletionTime: zod.string().nullish(),
  feedback: zod.string().nullish(),
  createdById: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
  assignees: zod.array(
    zod.object({
      id: zod.number(),
      email: zod.string(),
      name: zod.string(),
      role: zod.enum(["manager", "employee"]),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Get task count summary grouped by status
 */
export const GetTaskSummaryResponse = zod.object({
  todo: zod.number(),
  in_progress: zod.number(),
  done: zod.number(),
  total: zod.number(),
});