# lib/api-zod/src/generated/types/createTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTaskBody {
  title: string;
  description: string;
  /** @nullable */
  deadline?: string | null;
  assigneeIds?: number[];
}