# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: number;
  title: string;
  description: string;
  status: TaskStatus;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  completionPercent?: number | null;
  /** @nullable */
  expectedCompletionTime?: string | null;
  /** @nullable */
  feedback?: string | null;
  createdById: number;
  createdAt: string;
  updatedAt: string;
}