# lib/api-zod/src/generated/types/taskSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskSummary {
  todo: number;
  in_progress: number;
  done: number;
  total: number;
}