# lib/api-zod/src/generated/types/taskWithAssignees.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskWithAssigneesStatus } from "./taskWithAssigneesStatus";
import type { User } from "./user";

export interface TaskWithAssignees {
  id: number;
  title: string;
  description: string;
  status: TaskWithAssigneesStatus;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  completionPercent?: number | null;
  /** @nullable */
  expectedCompletionTime?: string | null;
  /** @nullable */
  feedback?: string | null;
  createdById: number;
  createdAt: string;
  updatedAt: string;
  assignees: User[];
}