# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyStatus } from "./updateTaskBodyStatus";

export interface UpdateTaskBody {
  status?: UpdateTaskBodyStatus;
  /** @nullable */
  completionPercent?: number | null;
  /** @nullable */
  expectedCompletionTime?: string | null;
  title?: string;
  description?: string;
  /** @nullable */
  deadline?: string | null;
  assigneeIds?: number[];
}